handle nextjs errors
